requirements.txt
# Core dependencies
fastapi>=0.110.2
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1

# LangChain (minimal for direct Gemini usage)
langchain-core>=0.1.13
langchain-community>=0.1.13

# Web and Async
aiofiles>=23.1.0
jinja2>=3.1.2
requests>=2.31.0

# Data Processing
beautifulsoup4>=4.12.2
lxml>=4.9.0
html5lib>=1.1

# Data Validation
pydantic>=2.0.0

# ASGI Server (for production)
gunicorn>=21.2.0

# Google Generative AI
google-generativeai>=0.8.0